split xgeom file to source code (extact_geom.py) and command code (command.py)
