+++ gaussianutility.egg-info/entry_points.txt
@@ -1,3 +1,3 @@
 [console_scripts]
-xgeom = gaussianutility.xgeom:main
+xgeom = gaussianutility.command:xgeom
