+++ setup.py
@@ -11,7 +11,7 @@ setuptools.setup(
     description="Useful utilities to use Gaussian09 software",
     long_description=long_description,
     long_description_content_type="text/markdown",
-    entry_points={'console_scripts': ['xgeom=gaussianutility.xgeom:xgeom']},
+    entry_points={'console_scripts': ['xgeom=gaussianutility.command:xgeom']},
     url="https://github.com/Sungil-Hong/gaussianutility",
     packages=setuptools.find_packages(),
     classifiers=[
