+++ test/3trimerPrism.SiAl.geom.com
@@ -1,0 +1,35 @@
+# opt freq=noraman 6-311g(d,p) scrf=(cpcm,solvent=water)  m062x temperature=373
+
+3trimerPrism.SiAl.geom.com
+
+0 1
+ Si   1.569482  -1.562712   0.811615
+  O   1.646550  -1.274805  -0.823861
+ Si   1.496638   0.075126  -1.757103
+  O  -0.018488   0.108541  -2.387862
+  O   0.111839  -2.204404   1.137380
+ Si  -1.528864   0.068469  -1.737418
+  O  -1.762830  -1.301566  -0.907909
+  O  -2.578754  -2.808909   1.481954
+  H  -3.535149  -2.742702   1.598914
+  O   2.773202  -2.577305   1.262513
+  H   2.663668  -3.507280   1.051400
+  O   2.481837   0.006822  -3.057499
+  H   3.428414   0.007178  -2.897176
+  O   1.755687   1.366778  -0.809178
+  O  -1.656809   1.342796  -0.699848
+  O  -2.631011   0.166130  -2.937711
+  H  -2.553890   0.892968  -3.560321
+  O  -0.092047   2.098200   1.328705
+ Si  -1.550477   1.497277   0.951178
+  O   2.607248   2.677669   1.672444
+  H   2.297680   3.475691   2.120618
+  O  -2.661318   2.574864   1.485415
+  H  -3.586248   2.346946   1.367738
+  O  -1.768065   0.005132   1.567851
+  O   1.790049  -0.135110   1.556134
+ Al   1.451865   1.441502   0.905962
+ Al  -1.438524  -1.515535   0.791231
+  H  -2.252185  -3.621626   1.890077
+  H   3.572643   2.658351   1.661219
+
